refactor(desktop): extract coding-agent thread model helpers
Move the pure thread-model helper layer out of the coding-agent-workspace
store into stores/coding-agent/thread-model.ts. Functions are moved verbatim
with zero behavior change; the store imports them. The seq-counter-mutating
clear*State helpers stay in the store because they are not pure.
